# api/openapi.yaml
openapi: 3.0.3
info:
  title: Rastreia API
  version: 0.1.0
  description: |
    Delivery tracking API. Carriers (transportadoras) sign up on their own and are tenants: every
    authenticated route only sees the caller's carrier, and anything from another carrier answers 404.
    Carriers create deliveries and drivers; drivers update status; customers follow a public tracking link.

    Every route counts toward a global limit of 120 requests per minute per IP, or per /64 network
    on IPv6 (429 with Retry-After). Bodies are JSON up to 1 MB; unknown fields and text holding
    `\u0000` answer 400. A request that runs for more than 15 seconds answers 504. Errors always
    come as `{"error": "..."}`, plus `fields` on 422.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    ValidationError:
      type: object
      required: [error, fields]
      properties:
        error: { type: string, example: invalid input }
        fields:
          type: object
          description: The message for each field that failed, in English (the front translates it)
          additionalProperties: { type: string }
          example: { postal_code: must have 8 digits }
    User:
      type: object
      required: [id, name, email, role, carrier_id, created_at]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        carrier_id: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
    Role:
      type: string
      enum: [carrier, driver]
      description: carrier runs a carrier (transportadora); driver delivers for it
    Carrier:
      type: object
      required: [id, name, document]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, example: Expresso Sul }
        document: { type: string, nullable: true, description: CNPJ without punctuation }
    Me:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [carrier]
          properties:
            carrier: { $ref: "#/components/schemas/Carrier" }
    Password:
      type: string
      minLength: 10
      description: At most 72 bytes. Common passwords and passwords made from the e-mail are refused.
    SignUp:
      type: object
      additionalProperties: false
      required: [carrier_name, name, email, password]
      properties:
        carrier_name: { type: string, maxLength: 120 }
        document:
          type: string
          nullable: true
          description: |
            Optional CNPJ, with or without punctuation. The alphanumeric format (letters in the first
            12 characters) is accepted. Must be unique.
        name: { type: string, maxLength: 120, description: Name of the person who runs the carrier }
        email: { type: string, format: email, maxLength: 254 }
        password: { $ref: "#/components/schemas/Password" }
    Summary:
      type: object
      required: [since, by_status, unassigned]
      properties:
        since: { type: string, format: date-time, description: Deliveries created from here on (30 days) are counted }
        by_status:
          type: object
          description: Every status is present, with 0 when there is none
          additionalProperties: { type: integer }
          example: { pending: 4, picked_up: 2, in_transit: 3, delivered: 20, failed: 1 }
        unassigned: { type: integer, description: "Pending deliveries without a driver, whenever created" }
    CreateDriver:
      type: object
      additionalProperties: false
      required: [name, email, password]
      properties:
        name: { type: string, maxLength: 120 }
        email: { type: string, format: email, maxLength: 254 }
        password: { $ref: "#/components/schemas/Password" }
...
    Delivery:
      type: object
      required:
        - id
        - tracking_code
        - recipient_name
        - recipient_email
        - address
        - recipient_phone
        - postal_code
        - street
        - number
        - complement
        - district
        - city
        - state
        - address_reference
        - latitude
        - longitude
        - status
        - driver_id
        - created_at
        - updated_at
        - completed_at
        - anonymized_at
      properties:
        id: { type: integer, format: int64 }
        tracking_code: { type: string, example: RS7K2M9QXA4P }
        recipient_name: { type: string }
        recipient_email: { type: string, description: "An e-mail; empty once anonymized" }
        address:
          type: string
          description: The whole address in one line, built from the parts below
          example: Praça da Sé, 10, Apto 2 - Sé, São Paulo - SP, 01001-000
        recipient_phone: { type: string, example: "11987654321", description: "Digits only; empty for deliveries created before the address had parts" }
        postal_code: { type: string, example: "01001000", description: "CEP, 8 digits; empty for deliveries created before the address had parts" }
        street: { type: string }
        number: { type: string, example: "S/N" }
        complement: { type: string, description: Empty when there is none }
        district: { type: string, description: Bairro }
        city: { type: string }
        state: { type: string, example: SP, description: UF }
        address_reference: { type: string, description: "Ponto de referência or notes for the driver; empty when there is none" }
        latitude: { type: number, format: double, minimum: -90, maximum: 90, nullable: true }
        longitude: { type: number, format: double, minimum: -180, maximum: 180, nullable: true }
        status: { $ref: "#/components/schemas/Status" }
        driver_id: { type: integer, format: int64, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: When the delivery last became delivered or failed
        anonymized_at:
          type: string
          format: date-time
          nullable: true
          description: |
            When the recipient's data was erased: RETENTION_DAYS after the delivery was finished, or a
            year after it was created if it never was. The name becomes "Destinatário removido"; e-mail,
            phone, address (all parts but city and state), reference, map position and event notes
            are emptied. Such a delivery answers 409 to PATCH and to new events, and its public link
            answers 404.
    Status:
      type: string
      enum: [pending, picked_up, in_transit, delivered, failed]
...
    Session:
      type: object
      required: [token, role, expires_in]
      properties:
        token: { type: string, description: Access token for the Authorization header }
        role: { $ref: "#/components/schemas/Role" }
        expires_in: { type: integer, description: Access token lifetime in seconds, example: 900 }
    Tracking:
      type: object
      description: "Public view of a delivery. No e-mail, phone, address, map position, last name, driver or notes."
      required: [tracking_code, carrier_name, status, recipient_first_name, updated_at, events]
      properties:
        tracking_code: { type: string, example: RS7K2M9QXA4P }
        carrier_name: { type: string, example: Expresso Sul }
        status: { $ref: "#/components/schemas/Status" }
        recipient_first_name: { type: string, example: Maria }
        updated_at: { type: string, format: date-time }
        events:
          type: array
          items:
            type: object
            required: [status, created_at]
            properties:
              status: { $ref: "#/components/schemas/Status" }
              created_at: { type: string, format: date-time }
...
    Route:
      type: object
      description: |
        The driver's route for today (Brazil's day). Packages to the same address (CEP, street and
        number) are one stop; packages are numbered 1..N in delivery order. A package the carrier
        gave to another driver leaves the route.
      required: [date, total_packages, stops]
      properties:
        date: { type: string, format: date }
        total_packages: { type: integer }
        stops:
          type: array
          items:
            type: object
            required: [number, address, latitude, longitude, packages]
            properties:
              number: { type: integer, example: 1 }
              address: { type: string, description: Without the complement }
              latitude: { type: number, format: double, nullable: true }
              longitude: { type: number, format: double, nullable: true }
              packages:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/Delivery"
                    - type: object
                      required: [position]
                      properties:
                        position: { type: integer, description: "Position of the package in the route, from 1" }
  responses:
    BadRequest:
      description: Malformed request (invalid JSON, empty body, unknown field or malformed id)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Missing or invalid token
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Role not allowed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    ForeignOrigin:
      description: The request came from a browser on a site not listed in CORS_ORIGINS
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    TooManyRequests:
      description: Too many requests or failed logins; wait for Retry-After seconds
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Invalid:
      description: Validation failed
      content: { application/json: { schema: { $ref: "#/components/schemas/ValidationError" } } }
  parameters:
...
    code:
      name: code
      in: path
      required: true
      schema: { type: string, example: RS7K2M9QXA4P }
...
paths:
...
  /auth/login:
    post:
      summary: Log in and get a short-lived JWT plus a refresh cookie
      description: |
        The access token lasts JWT_TTL (15 minutes by default). The refresh token comes in the
        rastreia_refresh cookie (HttpOnly, SameSite=Strict, Path=/auth) and is used on /auth/refresh.
        After 5 wrong passwords for an e-mail, it is locked for a while (429). A browser must send it
        from a site in CORS_ORIGINS; clients without an Origin header are fine.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Logged in. Sets the rastreia_refresh cookie.
          content: { application/json: { schema: { $ref: "#/components/schemas/Session" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Wrong e-mail or password (the same answer for both)
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "403": { $ref: "#/components/responses/ForeignOrigin" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503":
          description: "Login unavailable: the failed-login counter (Redis) did not answer"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /auth/signup:
    post:
      summary: Sign a carrier up and log its owner in
      description: |
        Creates the carrier and the account of the person who runs it, with the carrier role, and
        answers like /auth/login (access token plus the rastreia_refresh cookie). Has its own rate
        limit, the same size as login's (10 per minute per IP).
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SignUp" } } }
      responses:
        "200":
          description: Signed up and logged in. Sets the rastreia_refresh cookie.
          content: { application/json: { schema: { $ref: "#/components/schemas/Session" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/ForeignOrigin" }
        "409":
          description: E-mail or CNPJ already in use
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/refresh:
    post:
      summary: Trade the refresh cookie for a new access token and a new cookie
      description: |
        Each refresh token works once. Reusing one revokes the whole session. Requires an Origin
        listed in CORS_ORIGINS; the browser must send the request with credentials.
      parameters:
        - { name: rastreia_refresh, in: cookie, required: true, schema: { type: string } }
      responses:
        "200":
          description: New tokens. Sets a new rastreia_refresh cookie.
          content: { application/json: { schema: { $ref: "#/components/schemas/Session" } } }
        "401":
          description: Missing, expired, revoked or reused refresh token
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "403":
          description: Origin missing or not in CORS_ORIGINS
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /auth/logout:
    post:
      summary: Revoke the session and clear the refresh cookie
      description: Answers 204 even without a cookie. Requires an Origin listed in CORS_ORIGINS.
      parameters:
        - { name: rastreia_refresh, in: cookie, required: false, schema: { type: string } }
      responses:
        "204": { description: Logged out }
        "403":
          description: Origin missing or not in CORS_ORIGINS
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /me:
    get:
      summary: The logged-in user and their carrier
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Current user
          content: { application/json: { schema: { $ref: "#/components/schemas/Me" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /summary:
    get:
      summary: Carrier dashboard numbers
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Summary
          content: { application/json: { schema: { $ref: "#/components/schemas/Summary" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /drivers:
    get:
      summary: List the carrier's drivers
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Drivers
          content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/User" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      summary: Create a driver for the carrier
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateDriver" } } }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/User" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: E-mail already in use
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422": { $ref: "#/components/responses/Invalid" }
...
  /me/route:
    get:
      summary: The driver's route for today, empty until a package is scanned
      security: [{ bearer: [] }]
      responses:
        "200":
          description: The route
          content: { application/json: { schema: { $ref: "#/components/schemas/Route" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /me/route/order:
    put:
      summary: Save the order the driver chose
      description: Packages of the same stop stay together, in the place of the first of them.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [delivery_ids]
              properties:
                delivery_ids:
                  type: array
                  description: Every package of the route, once, in the new order
                  items: { type: integer, format: int64 }
      responses:
        "200":
          description: The route
          content: { application/json: { schema: { $ref: "#/components/schemas/Route" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Invalid" }
  /me/route/optimize:
    post:
      summary: Put the stops in a short order
      description: |
        Nearest neighbor and 2-opt over straight-line distances, starting near the driver's position
        when sent. Stops without a map position go to the end. Send {} to start from the first stop.
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                latitude: { type: number, format: double, minimum: -90, maximum: 90, nullable: true }
                longitude: { type: number, format: double, minimum: -180, maximum: 180, nullable: true }
      responses:
        "200":
          description: The route
          content: { application/json: { schema: { $ref: "#/components/schemas/Route" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Invalid" }
  /public/tracking/{code}:
    parameters: [{ $ref: "#/components/parameters/code" }]
    get:
      summary: Public tracking page data (no login)
      description: |
        Unknown, malformed, expired and anonymized codes answer 404. The link expires 30 days after
        the delivery is delivered or last failed. Limited to 30 requests per minute per IP.
      responses:
        "200":
          description: Tracking
          content: { application/json: { schema: { $ref: "#/components/schemas/Tracking" } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /public/push/key:
    get:
      summary: VAPID public key for Web Push (no login)
      description: Only served when the API has VAPID_PUBLIC_KEY. The browser passes it to PushManager.subscribe().
      responses:
        "200":
          description: Key
          content:
            application/json:
              schema:
                type: object
                required: [public_key]
                properties: { public_key: { type: string, description: base64url } }
        "404":
          description: Web Push is off (the API has no VAPID_PUBLIC_KEY)
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /public/tracking/{code}/push:
    parameters: [{ $ref: "#/components/parameters/code" }]
    post:
      summary: Get push notifications about this delivery on this browser (no login)
      description: |
        The body is the browser's PushSubscription.toJSON(). The endpoint must be https on a known
        browser push service (Google, Mozilla, Microsoft, Apple), since the worker posts to it.
        Up to 10 browsers per delivery (one already subscribed can always subscribe again); delivered
        deliveries answer 409. Subscriptions are deleted when the delivery is delivered or anonymized.
        Shares the tracking rate limit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [endpoint, keys]
              properties:
                endpoint: { type: string, example: "https://fcm.googleapis.com/fcm/send/abc" }
                expirationTime: { type: number, nullable: true, description: Sent by browsers; ignored }
                keys:
                  type: object
                  additionalProperties: false
                  required: [p256dh, auth]
                  properties: { p256dh: { type: string }, auth: { type: string } }
      responses:
        "204": { description: Subscribed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Delivered, or too many browsers"
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
    delete:
      summary: Stop push notifications about this delivery on this browser (no login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [endpoint]
              properties: { endpoint: { type: string } }
      responses:
        "204": { description: Unsubscribed (also when it was not subscribed) }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...