# internal/store/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package store

import (
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type Carrier struct {
	ID        int64
	Name      string
	Document  *string
	CreatedAt time.Time
}

type Delivery struct {
	ID               int64
	TrackingCode     string
	RecipientName    string
	RecipientEmail   string
	Address          string
	Status           string
	DriverID         *int64
	CreatedAt        time.Time
	UpdatedAt        time.Time
	CompletedAt      *time.Time
	AnonymizedAt     *time.Time
	CarrierID        int64
	RecipientPhone   string
	PostalCode       string
	Street           string
	Number           string
	Complement       string
	District         string
	City             string
	State            string
	AddressReference string
	Latitude         *float64
	Longitude        *float64
}

type DeliveryEvent struct {
	ID          int64
	DeliveryID  int64
	Status      string
	Note        *string
	CreatedBy   *int64
	CreatedAt   time.Time
	PublishedAt *time.Time
}

type IdempotencyKey struct {
	UserID      int64
	Key         string
	RequestHash string
	DeliveryID  *int64
	CreatedAt   time.Time
}

type PushSubscription struct {
	ID         int64
	DeliveryID int64
	Endpoint   string
	P256dh     string
	Auth       string
	CreatedAt  time.Time
}

type RefreshToken struct {
	ID        int64
	UserID    int64
	FamilyID  string
	TokenHash []byte
	ExpiresAt time.Time
	UsedAt    *time.Time
	RevokedAt *time.Time
	CreatedAt time.Time
}

type Route struct {
	ID        int64
	CarrierID int64
	DriverID  int64
	RouteDate pgtype.Date
	CreatedAt time.Time
}

type RouteItem struct {
	RouteID    int64
	DeliveryID int64
	Position   int32
	AddedAt    time.Time
}

type User struct {
	ID           int64
	Name         string
	Email        string
	PasswordHash string
	Role         string
	CreatedAt    time.Time
	CarrierID    int64
}